Chatting_Platform/requirements.txt
Flask>=3.0.0
Flask-SQLAlchemy>=3.1.1
nats-py>=2.7.0
Flask-CORS>=4.0.0
Flask-SocketIO>=5.3.5
python-socketio>=5.10.0
python-engineio>=4.8.0
python-dotenv>=1.0.1
boto3>=1.26.0
python-jose[cryptography]>=3.3.0
requests>=2.31.0
SQLAlchemy==2.0.48
Werkzeug>=3.0.1
gunicorn>=21.2.0
python-socketio[client]>=5.10.0
